// src/app/core/features/dashboard/user-management/store/types.ts
export enum UserTypes  {
    GET_USERS = '[USER-MG] GET_USERS',
    GET_USERS_SUCCESS = '[USER-MG] GET_USERS_SUCCESS',
    GET_USER_BY_ID = '[USER-MG] GET_USER_BY_ID',
    GET_USER_BY_ID_SUCCESS = '[USER-MG] GET_USER_BY_ID_SUCCESS',
    ADD_USER = '[USER-MG] ADD_USER',
    ADD_USER_SUCCESS = '[USER-MG] ADD_USER_SUCCESS',
    CREATE_ROLE_SUCCESS = '[USER-MG] CREATE_ROLE_SUCCESS',
    UPDATE_USER = '[USER-MG] UPDATE_USER',
    UPDATE_USER_SUCCESS = '[USER-MG] UPDATE_USER_SUCCESS',
    DELETE_USER = '[USER-MG] DELETE_USER',
    DELETE_USER_SUCCESS = '[USER-MG] DELETE_USER_SUCCESS',
    GET_ROLES = '[USER-MG] GET_ROLES',
    ADD_ROLE = '[USER-MG] ADD_ROLE',
    UPDATE_ROLE = '[USER-MG] UPDATE_ROLE',
    GET_ROLES_SUCCESS = '[USER-MG] GET_ROLES_SUCCESS',
    ADD_ROLE_SUCCESS = '[USER-MG] ADD_ROLE_SUCCESS',
    UPDATE_ROLE_SUCCESS = '[USER-MG] UPDATE_ROLE_SUCCESS',
    SELECT_USER = '[USER-MG] SELECT_USER',
    THROW_ERROR = '[USER-MG] THROW_ERROR'
};